requirements.txt
altgraph==0.17
bitstring==3.1.7
et-xmlfile==1.1.0
future==0.18.2
numpy==1.25.1
openpyxl==3.1.2
pandas==1.1.5
pefile==2019.4.18
Pillow==8.1.2
psutil==5.8.0
pygame==2.1.2
pygame-menu==4.3.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pylink-square==0.8.2
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
six==1.15.0
typing_extensions==4.4.0
tzdata==2023.3
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
